// apps/web/src/extensions/slash-command/types.ts
import { Editor } from "@tiptap/core";

import { icons } from "lucide-react";

export interface Group {
  name: string;
  title: string;
  icon: React.ReactNode;
  commands: Command[];
}

export interface Command {
  name: string;
  label: string;
  description: string;
  aliases?: string[];
  iconName: keyof typeof icons;
  icon: React.ReactNode;
  action: (editor: Editor) => void;
  shouldBeHidden?: (editor: Editor) => boolean;
}

export interface MenuListProps {
  editor: Editor;
  items: Group[];
  command: (command: Command) => void;
}